// apps/desktop/src/common/models/types.ts
export enum ReadingDirection {
  LeftToRight = 'LeftToRight',
  RightToLeft = 'RightToLeft',
}

export enum PageStyle {
  Single = 'Single',
  Double = 'Double',
  LongStrip = 'LongStrip',
}

export enum OffsetPages {
  All = 'All',
  First = 'First',
  None = 'None',
}